Extract site settings link list
