+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.67",
-  "tag": "v3.4.67",
+  "version": "3.4.68",
+  "tag": "v3.4.68",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.66 <3.4.67"
+      ">=3.4.67 <3.4.68"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.66 first."
+    "message": "Update to v3.4.67 first."
   }
 }
